src/main/java/org/springframework/samples/petclinic/usuario/UsuarioService.java
package org.springframework.samples.petclinic.usuario;

import java.util.Optional;


import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.dao.DataAccessException;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;


@Service
public class UsuarioService {
    private UsuarioRepository usuarioRepository;

    @Autowired
    public UsuarioService(UsuarioRepository usuarioRepository){
        this.usuarioRepository=usuarioRepository;
    }

    public Optional<Usuario> findUserByNombreUsuario(String nombreUsuario){
        return usuarioRepository.findByNombreUsuario(nombreUsuario);
    }
    @Transactional
	public void saveUser(Usuario user) throws DataAccessException {
		user.setEnabled(true);
		usuarioRepository.save(user);
	}
}
